Added the ability to dynamically display a list of products in a dialog style div that is fixed position centered on the page.
In index.html you will see a new div that is a placeholder for display the products when a category is clicked.

In style.css you will this div is styled and also initially hidden by display:none.  Needed to add position relative to
html and body which we can then position the category-products div fixed in the middle of the browser when displayed.

In script.js we hooked up a javascript function to the a href click which references the displayProducts function.  This function takes in the key of the current category in the list.
When the a is clicked the products tied to the key are displayed in the category-products div and then the div is displayed.